requirements.txt
requests>=2.28.0
pandas>=1.5.0
google-cloud-bigquery>=3.0.0